added a reporter that can end simulation early and moved reporters to miner_utlis
